userkeyword js파일들 컨벤션에 따른 분리
